+++ app/src/main/java/jp/developer/retia/frozenword/ui/habbits/HabbitsActivity.kt
@@ -74,74 +74,3 @@ class HabbitsActivity : ComponentActivity() {
     }
 }
 
-@Preview
-@Composable
-fun PreviewHabbits() {
-    FrozenWordTheme {
-        Surface {
-            Habbits(
-                listOf(
-                    HabbitAndLog(Habbit(title = "hogehoge", trigger = "まるまるしたら"), emptyList()),
-                    HabbitAndLog(Habbit(title = "hogehoge", trigger = "まるまるしたら"), emptyList()),
-                    HabbitAndLog(Habbit(title = "hogehoge", trigger = "まるまるしたら"), emptyList()),
-                )
-            )
-        }
-    }
-}
-
-@OptIn(ExperimentalFoundationApi::class)
-@Composable
-fun Habbits(
-    habbits: List<HabbitAndLog>
-) {
-    LazyColumn(modifier = Modifier.background(Color.Gray).wrapContentHeight()) {
-        items(habbits) { habbit ->
-            Card(modifier = Modifier.padding(2.dp)) {
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .wrapContentHeight()
-                        .padding(8.dp)
-                ) {
-                    Column(modifier = Modifier.weight(1f)) {
-                        Text(habbit.habbit.trigger, fontSize = 13.sp)
-                        Text(habbit.habbit.title, fontSize = 16.sp)
-                    }
-                    TaskDots(modifier = Modifier.weight(1f), 24)
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun TaskDots(modifier: Modifier = Modifier, days: Int) {
-    val columnSize = min(days / 7, 4)
-    Column {
-        repeat(columnSize) {
-            Row {
-                repeat(7) {
-                    Row(
-                        modifier = Modifier
-                            .width(12.dp)
-                            .height(12.dp)
-                            .background(Color.Blue),
-                    ) {}
-                }
-            }
-        }
-        if (columnSize < 4) {
-            Row {
-                repeat(days % 7) {
-                    Row(
-                        modifier = Modifier
-                            .width(12.dp)
-                            .height(12.dp)
-                            .background(Color.Blue)
-                    ) {}
-                }
-            }
-        }
-    }
-}
